backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-appconf==1.0.6
django-measurement==3.2.4
measurement==3.2.2
mpmath==1.3.0
sqlparse==0.5.1
sympy==1.13.3
djangorestframework
djangorestframework-simplejwt
django-cors-headers

